FIle location and calculation changes
Changed location of files in server_func to be sent to share_to_send instead of share_received.

Implemented error handling in the calculation functions in server_func.

Clarified README files of the three folders for a better understanding of what the files parameters are meant to be and what files are located in the folder.